backend/requirements.txt
# FastAPI Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlmodel==0.0.22
alembic==1.14.0
pyodbc==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.27.0

# Email validation
email-validator==2.2.0
sqlmodel

# Google OAuth
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Two-Factor Authentication (2FA)
pyotp==2.9.0
qrcode[pil]==7.4.2